% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjgame.R
\name{rjgame}
\alias{rjgame}
\title{Create a simple javascript game as a htmlwidget}
\usage{
rjgame(message = "", width = NULL, height = NULL, elementId = NULL)
}
\description{
Use the arrowkeys to move the red square and avoid the obstacles.
The move you avoid, the higher the score.
}
\examples{
rjgame()
}
